20210512-01:31

+++ client/src/Components/Person/Person.component.jsx
@@ -1,16 +1,22 @@
-import React from "react";
+import React, { useState } from "react";
 import "./Person.component.css";
 import api from "../../API/api";
 
-const Person = ({ data, remove }) => {
+const Person = ({ data, remove, userName }) => {
+  const [enableDisable, setEnableDisable] = useState("person-item-button");
   // ----------- delete item ----------------------------
   const handleDeleteItem = async () => {
-    const token = await localStorage.getItem("token");
-    const response = await api.delete(`/persons/${data._id}`, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    console.log(response);
-    remove(data._id);
+    if (data.name !== userName) {
+      setEnableDisable("person-item-button");
+      const token = await localStorage.getItem("token");
+      const response = await api.delete(`/persons/${data._id}`, {
+        headers: { Authorization: `Bearer ${token}` },
+      });
+      console.log(response);
+      remove(data._id);
+    } else {
+      setEnableDisable("person-item-button-disabled");
+    }
   };
   // ------------------------------------------------------
   return (
@@ -25,7 +31,7 @@ const Person = ({ data, remove }) => {
       </div>
       <div className="person-item-buttons">
         <div className="person-item-button">Edit</div>
-        <div onClick={handleDeleteItem} className="person-item-button">
+        <div onClick={handleDeleteItem} className={enableDisable}>
           Delete
         </div>
       </div>
